File version tagging and better content reservation handling (#33)
* new possibilities for variable handling

* add the possibility to create 4byte signatures

* several improvements to make the code more readable

* implement file size reservations

* improve the way to find the newest content version to prepare for version restoring

* implement version tagging

* add space reservation for new content

* fixing a bug in the interpretation of reserved blocks

* fixing several bugs that emerged during programming

* removing an obsolete methode

* add a tool to fix files in the console

* add tag and fix command to the console

* improve the output of env in the console

* Make the files system more robust and mark a memory leak

* cleanup AI slop

* fix the bug where an unfinished file content would be marked as finished

+++ app/src/Flash/flash.cpp
@@ -96,7 +96,7 @@ int Flash::write(const void* buffer, size_t size, const void* address) {
   }
 
   int page_count = FlashHAL::calculatePage(size);
-  if (FlashHAL::calculatePageAddress(page_count) != (int)size) {
+  if (FlashHAL::calculatePageAddress(page_count) != size) {
     return -3; // Buffer size not aligned to page size
   }
 
@@ -116,7 +116,7 @@ int Flash::erase(const void* address, int length) {
     return -2; // Address not aligned to sector start
   }
 
-  if (FlashHAL::calculateSectorAddress(sector_count) != length) {
+  if (FlashHAL::calculateSectorAddress(sector_count) != (size_t)length) {
     return -3; // Erase length not aligned to sector size
   }
